requirements.txt
fastapi
loguru
uvicorn
pydantic-ai
pandas
pyaml
pillow
matplotlib
streamlit
st_pages